Move PipelineResult and PipelineError into separate module
